Pulled app level functions to util file.
Cleaning up components.
