+++ src/client/components/App.jsx
@@ -1,4 +1,6 @@
 import React, {Component} from 'react';
+import { handleSelectedStation, handlePlayState, seekStation, generateRandomStationId, loadStations } from '../util/ClientFunctions.jsx';
+
 import AudioPlayer from './AudioPlayer.jsx';
 import StationList from './StationList.jsx';
 
@@ -14,90 +16,15 @@ class App extends Component {
         type: ""
       }
     },
-    this.handlePlayState = this.handlePlayState.bind(this);
-    this.handleSelectedStation = this.handleSelectedStation.bind(this);
-    this.seekStation = this.seekStation.bind(this);
-    this.generateRandomStationId = this.generateRandomStationId.bind(this);
-  }
-
-  handleSelectedStation(id, name, stream, type) {
-    this.setState({
-      selectedStation: {
-        id: id,
-        name: name,
-        stream: stream,
-        type: type
-      }
-    });
-    this.handlePlayState();
-  }
-
-  handlePlayState() {
-    let player = document.getElementById("player");
-    document.getElementById("player").load();
-
-    // Displays the loading icon while the media is loading.
-    player.addEventListener("loadstart", function() {
-      pButton.className = "fa fa-spinner fa-pulse fa-2x fa-fw";
-    });
-
-    // Plays the audio when it is ready.
-    player.addEventListener("canplaythrough", function() {
-      document.getElementById("player").play();
-    });
-
-    // Changes the loading icon to a pause icon.
-    player.addEventListener("playing", function() {
-      pButton.className = "";
-      pButton.className = "fa fa-pause fa-2x";
-    });
-  }
-
-  generateRandomStationId() {
-    const seekLength = this.state.stations.length - 1;
-    return Math.floor(Math.random() * seekLength);
-  }
-
-  seekStation() {
-    const randomStationId = this.generateRandomStationId();
-
-    for(const station of this.state.stations) {
-
-      // Checks to see if the seek matches the currently playing station.
-      if(randomStationId === this.state.selectedStation.id) {
-        this.seekStation();
-        return;
-
-      // Plays the seeked station.
-      } else if(randomStationId === station.id) {
-          this.handleSelectedStation(station.id, station.name, station.audio_feed, station.stream_type);
-      }
-    }
+    this.handlePlayState = handlePlayState.bind(this);
+    this.handleSelectedStation = handleSelectedStation.bind(this);
+    this.seekStation = seekStation.bind(this);
+    this.generateRandomStationId = generateRandomStationId.bind(this);
+    this.loadStations = loadStations.bind(this)
   }
 
   componentDidMount() {
-
-    function shuffle(sourceArray) {
-      for (var i = 0; i < sourceArray.length - 1; i++) {
-        var j = i + Math.floor(Math.random() * (sourceArray.length - i));
-
-        var temp = sourceArray[j];
-        sourceArray[j] = sourceArray[i];
-        sourceArray[i] = temp;
-      }
-      return sourceArray;
-    }
-
-    fetch('/api/stations')
-      .then((response) => {
-        return response.json()
-      }).then((json) => {
-        this.setState({ stations: shuffle(json) })
-        console.log('parsed json', json)
-      }).catch((ex) => {
-        console.log('parsing failed', ex)
-      });
-    console.log("Mounted");
+    this.loadStations();
   }
 
   render() {
