+++ src/client/util/ClientFunctions.jsx
@@ -1,5 +1,60 @@
 import React, { Component } from 'react';
 
+export const handleSelectedStation = function(id, name, stream, type) {
+    this.setState({
+        selectedStation: {
+            id: id,
+            name: name,
+            stream: stream,
+            type: type
+        }
+    });
+    this.handlePlayState();
+}
+
+export const generateRandomStationId = function() {
+    const seekLength = this.state.stations.length - 1;
+    return Math.floor(Math.random() * seekLength);
+}
+
+export const seekStation = function() {
+    const randomStationId = this.generateRandomStationId();
+
+    for (const station of this.state.stations) {
+
+        // Checks to see if the seek matches the currently playing station.
+        if (randomStationId === this.state.selectedStation.id) {
+            this.seekStation();
+            return;
+
+            // Plays the seeked station.
+        } else if (randomStationId === station.id) {
+            this.handleSelectedStation(station.id, station.name, station.audio_feed, station.stream_type);
+        }
+    }
+}
+
+export const handlePlayState = function() {
+    let player = document.getElementById("player");
+    document.getElementById("player").load();
+
+    // Displays the loading icon while the media is loading.
+    player.addEventListener("loadstart", function () {
+        pButton.className = "fa fa-spinner fa-pulse fa-2x fa-fw";
+    });
+
+    // Plays the audio when it is ready.
+    player.addEventListener("canplaythrough", function () {
+        document.getElementById("player").play();
+    });
+
+    // Changes the loading icon to a pause icon.
+    player.addEventListener("playing", function () {
+        pButton.className = "";
+        pButton.className = "fa fa-pause fa-2x";
+    });
+}
+
 // Function to take a station div selection and pass it to the app level as current player.
 export const onStationSelect = function(event) {
     this.props.handleSelectedStation(this.state.name, this.state.stream, this.state.type);
@@ -41,5 +96,28 @@ export const makeItPlay = function() {
         pButton.className = "";
         pButton.className = "fa fa-play fa-2x";
     }
+}
+
+export const shuffle = function (sourceArray) {
+    for (var i = 0; i < sourceArray.length - 1; i++) {
+        var j = i + Math.floor(Math.random() * (sourceArray.length - i));
+
+        var temp = sourceArray[j];
+        sourceArray[j] = sourceArray[i];
+        sourceArray[i] = temp;
+    }
+    return sourceArray;
+}
+
+export const loadStations = function() {
+    fetch('/api/stations')
+        .then((response) => {
+            return response.json()
+        }).then((json) => {
+            this.setState({ stations: shuffle(json) })
+            console.log('parsed json', json)
+        }).catch((ex) => {
+            console.log('parsing failed', ex)
+        });
 }
